get all things down

+++ resources/views/tasks/index.blade.php
@@ -3,26 +3,64 @@
 @section('content')
 
     <div class="container">
-        <h1>大家的需求</h1>
+        @include('partial.createTask')
         @foreach($tasks as $task)
             <div class="row">
                     <div class="well well-lg col-md-12">
                         <p class="text-primary text-left lead">{{ $task->content }}</p>
 
                         <div class="row">
-                            <div class="col-md-1 col-md-offset-9">
-                                <div class="btn btn-primary">{{ $task->created_at->diffForHumans() }}</div>
+
+                            <div class="col-md-1">
+
+                                @if($task->isProcessing())
+                                    <div class="btn btn-success">{{ $task->receiver->name }} 正在帮忙</div>
+                                @endif
+
+                                @if($task->isConfirmed())
+                                    <div class="btn btn-info">{{ $task->receiver->name }} 帮的忙</div>
+                                @endif
+
+                            </div>
+
+
+                            <div class="col-md-1 col-md-offset-1">
+                                <div class="btn btn-info">{{ $task->created_at->diffForHumans() }}</div>
                             </div>
 
                             <div class="col-md-1">
-                                <div class="btn btn-info">{{ $task->sender->name }}</div>
+                                <div class="btn btn-info">求助人：{{ $task->sender->name }}</div>
                             </div>
 
-                            <div class="col-md-1">
-                                {!! Form::open(['method' => 'DELETE','url' => ['tasks',$task->id]])  !!}
+                            <div class="col-md-1 col-md-offset-1">
+
+                                @if($task->isNotReceived() && (Auth::guest() || $task->notSendByUser(Auth::user())))
+                                    <a class="btn btn-success" href="{{ url('tasks/take').'/'.$task->id }}">我要帮忙</a>
+                                @endif
+
+                                @if(Auth::check() && $task->isCanDelete(Auth::user()))
+                                    {!! Form::open(['method' => 'DELETE','url' => ['tasks',$task->id]]) !!}
                                     {!! Form::submit('删除',['class' => 'btn btn-danger']) !!}
-                                {!! Form::close() !!}
+                                    {!! Form::close() !!}
+                                @endif
+
+                                @if(Auth::check() && $task->isCanConfirm(Auth::user()))
+                                    <a class="btn btn-warning" href="{{ url('tasks/confirm').'/'.$task->id }}">确认完成</a>
+                                @endif
+
+                                @if($task->isConfirmed())
+                                    <div class="btn btn-info">已完成</div>
+                                @endif
+
                             </div>
+
+                            <div class="col-md-1">
+
+                                @if(Auth::check() && $task->isCanRemove(Auth::user()))
+                                    <a class="btn btn-warning" href="{{ url('tasks/remove').'/'.$task->id }}">取消{{ $task->receiver->name }}的帮助</a>
+                                @endif
+                            </div>
+
 
                         </div>
